backend/hang_backend/requirements.txt
django==4.1.3
django-rest-framework
dj-rest-auth==2.2.5
django-cors-headers
django-rest-knox
boto3==1.26.9
channels==3.0.5
django-allauth==0.51.0
daphne==3.0.2
emoji
python-dateutil
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
uvicorn[standard]
python-dotenv
